refactor(releases): move useReleasesStore out of legacy
