+++ packages/sanity/src/core/releases/contexts/ReleasesMetadataProvider.tsx
@@ -2,9 +2,9 @@ import {useCallback, useContext, useEffect, useMemo, useState} from 'react'
 import {useObservable} from 'react-rx'
 import {ReleasesMetadataContext} from 'sanity/_singletons'
 
-import {useReleasesStore} from '../index'
 import {type MetadataWrapper} from '../store/createReleaseMetadataAggregator'
 import {type ReleasesMetadata} from '../store/useReleasesMetadata'
+import {useReleasesStore} from '../store/useReleasesStore'
 
 /**
  * @internal
